refactor: piece 와, game 을 패키지로 분리
