+++ src/main/java/chess/domain/piece/Piece.java
@@ -1,7 +1,7 @@
-package chess.domain;
+package chess.domain.piece;
 
 import chess.domain.exception.IllegalPieceMoveException;
-import chess.domain.state.MoveState;
+import chess.domain.piece.state.MoveState;
 
 public class Piece {
