+++ src/main/java/chess/domain/piece/state/MovedPawnState.java
@@ -1,6 +1,6 @@
-package chess.domain.state;
+package chess.domain.piece.state;
 
-import chess.domain.ColorCompareResult;
+import chess.domain.piece.ColorCompareResult;
 
 public class MovedPawnState implements MoveState {
