+++ src/test/java/chess/domain/BoardTest.java
@@ -12,6 +12,9 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import chess.domain.exception.IllegalPieceMoveException;
+import chess.domain.game.Board;
+import chess.domain.piece.Piece;
+import chess.domain.piece.PieceType;
 import java.util.stream.Collectors;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
